Disable line buffering if the file is opened in binary mode; Also fix a couple of fopen/fdopen bugs
git-svn-id: https://nuttx.svn.sourceforge.net/svnroot/nuttx/trunk@4630 7fd9a85b-ad96-42d3-883c-3090e2eb8679

+++ nuttx/fs/fs_fdopen.c
@@ -237,7 +237,7 @@ FAR struct file_struct *fs_fdopen(int fd, int oflags, FAR _TCB *tcb)
            */
 
           stream->fs_filedes = fd;
-          stream->fs_oflags  = oflags;
+          stream->fs_oflags  = (uint16_t)oflags;
 
           sem_post(&slist->sl_sem);
           return stream;
